requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.2.0
motor==2.5.1
mypy-extensions==1.0.0
pathspec==0.11.1
platformdirs==3.5.3
psutil==5.9.1
pymongo==3.13.0
python-dotenv==0.20.0
pytz==2023.3
